+++ package/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.9.0
+
+### Breaking Changes
+
+- Status code changed in `RequestTimeLimit` layer:
+    - `408` -> `504`
+
+### What's Changed
+
+- Update documentation
+
 ## 0.8.0 (2025-09-06)
 
 ### Breaking Changes